+++ resources/assets/js/project.js
@@ -66,6 +66,18 @@ new Vue({
         .then(response => {
           console.log(response)
           this.projects = response.data
+          this.projects.forEach(project => {
+            axios.get('/project/hasMembers', {
+              params: {
+                prj_no: project.prj_no
+              }
+            }).then(response2 => {
+              project.hasMembers = response2.data.hasMembers
+            })
+            .catch(error2 => {
+              console.log(error2)
+            })
+          })
           this.filtered = this.projects
         })
         .catch(error => {
@@ -96,9 +108,32 @@ new Vue({
         })
     },
     destroy(index) {
-      axios.delete('/project/destroy', {
-        params: {
-          prj_no: this.projects[index].prj_no
+      bootbox.confirm({
+        title: 'Delete confirmation',
+        message: 'Do you really want to cancel this project ?',
+        buttons: {
+          cancel: {
+            label: 'No'
+          },
+          confirm: {
+            label: 'Yes'
+          }
+        },
+        callback: (confirm) => {
+          if(confirm) {
+            axios.delete('/project/destroy', {
+              params: {
+                prj_no: this.projects[index].prj_no
+              }
+            })
+            .then(response => {
+              console.log(response)
+              this.projects.splice(index, 1)
+            })
+            .catch(error => {
+              console.log(error)
+            })
+          }
         }
       })
         .then(response => {
